♻️ (#417) Board 상수 구조 개선
