--- src/features/board/constants/boardConstants.ts
@@ -1,5 +1,0 @@
-export const BOARD_KEYS = {
-  MY_TASKS: 'myTasks',
-  PROJECT_STATUS: 'projectStatus',
-  PROJECT_MEMBER: 'projectMember',
-} as const;
