--- src/features/board/constants/scroll.ts
@@ -1,1 +1,0 @@
-export const COLLAPSIBLE_SCROLL_THRESHOLD = 100;
